requirements.txt
# LangGraph agent
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-groq>=0.2.0

# Vector DB & embeddings
pinecone-client>=5.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# LLM client
groq>=0.4.0

# Data sources (no keys required)
yfinance>=0.2.0

# News (free tier key required)
newsapi-python>=0.2.7

# Backend
fastapi>=0.115.0
slowapi>=0.1.9

uvicorn>=0.30.0

# Frontend
dash>=2.18.0
dash-bootstrap-components>=1.6.0
plotly>=5.24.0

# Data transformation
dbt-core>=1.7.0
dbt-postgres>=1.7.0

# Caching (Upstash serverless Redis)
upstash-redis>=1.0.0
redis>=5.0.0

# Data models & utilities
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
